+++ testing/helpers/dataGridHelper.js
@@ -1,28 +1,8 @@
-import DataGrid from 'ui/data_grid';
-
-import $ from 'jquery';
-
 import 'generic_light.css!';
 
-DataGrid.defaultOptions({
-    options: {
-        loadingTimeout: 0
-    }
-});
+import { baseModuleConfig, createDataGrid } from './baseDataGridHelper.js';
 
-export const baseModuleConfig = {
-    beforeEach: function() {
-        this.clock = sinon.useFakeTimers();
-    },
-    afterEach: function() {
-        this.clock.restore();
-    }
-};
-
-export const createDataGrid = (options, $container) => {
-    const dataGridElement = ($container || $('#dataGrid')).dxDataGrid(options);
-
-    QUnit.assert.ok(dataGridElement);
-    const dataGrid = dataGridElement.dxDataGrid('instance');
-    return dataGrid;
+export {
+    baseModuleConfig,
+    createDataGrid
 };
