+++ testing/tests/DevExpress.ui.widgets.dataGrid/materialTheme.tests.js
@@ -1,7 +1,8 @@
-import 'material_blue_light.css!';
-
 import $ from 'jquery';
-import { createDataGrid, baseModuleConfig } from '../../helpers/dataGridHelper.js';
+
+import { createDataGrid, baseModuleConfig } from '../../helpers/baseDataGridHelper.js';
+
+import 'material_blue_light.css!';
 
 QUnit.testStart(function() {
     const markup = '<div id="dataGrid"></div>';
@@ -30,5 +31,35 @@ QUnit.module('Material theme', baseModuleConfig, () => {
 
         assert.equal(rightPadding, '16px');
         assert.equal(leftPadding, '16px');
+    });
+
+    QUnit.test('Vertical scroll does not bounce on updateDimensions (T1050513)', function(assert) {
+        const dataGridInstance = createDataGrid({
+            dataSource: Array(50).fill(0).map((_, i) => ({ id: i })),
+            columnHidingEnabled: true,
+            height: 702,
+            keyExpr: 'id',
+            paging: {
+                enabled: false
+            },
+            scrolling: {
+                useNative: false
+            }
+        });
+
+        this.clock.tick();
+
+        const scrollable = dataGridInstance.getScrollable();
+
+        // scroll to the end
+        scrollable.scrollTo(10000);
+
+        const beforeUpdateScrollPosition = scrollable.scrollTop();
+
+        dataGridInstance.updateDimensions();
+
+        const afterUpdateScrollPosition = scrollable.scrollTop();
+
+        assert.roughEqual(beforeUpdateScrollPosition, afterUpdateScrollPosition, 1, 'Scroll position is not change');
     });
 });
